Expanded SceneRenderer, moved Shader to a use/unuse system
Each material is rendered individually in SceneRenderer. Currently it is just doing the same thing it was doing before, rendering every material in the order we have it, but optimisations will change that. I added a basic list of things to implement. Also moved Shader to be a use/unuse system. Probably harder for an average person to understand but I am fine with it. The constant binding and unbinding of the shader in the uniform uploads was limiting performance once we were uploading uniforms outside of the shader (like the textures for each material) so it is better this way. Exciting times.