Materials now render independent of their model
The model aspect has been removed and we now just look at all the materials. Haven't implemented for shadows yet. Also made it so that opaque textures try to render front to back and transparent back to front. In reality it doesn't always work that well with materials being over an area and overlapping/being inside of other materials. It actually runs slower after implementing that possibly because of the added CPU time to sort the 2 vectors, so might just remove it. Next frustum culling which should improve performance. Been thinking about dynamic near/far clip planes and it would improve fighting Z a lot visually but it would mean having to compute the view matrix and upload it for every material whereas now we are only doing it once at the start. So I don't know.