Added frustum culling
Also made a new depth only shader that doesn't take texture so if a model is marked as opaque we can render it faster. Frustum culling works by seeing if a materials bounds are in our frustum. The track being so big every material was in the frustum, so I split the material up if it is bigger than a specified size. This way we have more materials but they cover a smaller area and we can actually benefit from frustum culling. It does suck that there are more materials though and it puts more load on the cpu to go through the extra materials.