Do depth pre pass
Instead of occlusion culling (seems complicated) I tried doing a depth pre pass and using that as the depth buffer, so the fragment shader won't try to shade fragments that are going to be covered by something else later. Was a pain and actually lost 0.3ms of rendering, maybe it didn't save time because I already sort the materials front to back so there wasn't much over draw anyway. But depth pre pass can be used for other things so it's good to just have it.