Fixed front to back rendering
And a lot more. Used RenderDoc to look at my scene. Had to use a modern core context for RenderDoc to work, that broke my images because it turns out I wasn't actually creating the VAO it would just stay at the default 0 which was fine in previous versions. Noticed geometry was rendering back to front, so I fixed it in the code. Also saw that with a depth pass it was taking just as long (which I knew already) so I disabled the depth pass and we render a lot faster. Occlusion culling would be nice but it seems like a big task, proper sorting of front to back means that there isn't as much overdraw so it minimises the problem, but the vertex shader is still called lots of times. Also started on implementing IBL properly. Started on generating the BRDF (which I had as an image in assets but I will generate it myself) and will generate an irradiance map and a pre filtered environment map. Came from when I remembered that I skipped this step and just got something that looked fine. Trying to read the BRDFLUT properly now leaves the image very bright.