Implemented IBL properly
Long day. Implemented RenderTexture cube maps, so that we could create the irradiance and pre filtered environment maps of the skybox to be used for the IBL. Also computed my own BRDF LUT and I can now sample from it properly. It is very hard to tell the difference from what it was before. Theoretically this could lead to better IBL with a system for a "pre rendered environment and irradiance maps" in certain areas (such as indoors) can have pre rendered maps and we can switch to those when in those areas. But as far as I am concerned right now it looks the same as before.