Draw shading to a render texture
Did lots more but ultimately that's all. Shading is done to a render texture so that we have access to it for post processing. Currently just blit it to the colour framebuffer but will change that. Tried to get the depth buffer to work again but couldn't, gave up so we now just render the depth separately and don't bind it to the depth buffer. Theoretically this way is more overdraw and the vertices have to be processed twice but at least it works. Trouble is the anti aliasing, specifically MSAA as it has to match the SDL window and the render texture which gets complicated. Currently hardcoded. Next I guess rendering to the screen properly using a quad and then post processing, like AO. Also commented out everything and added the Sponza model to the scene for visual testing.