Implemented PCSS
Also re-wrote the shadow map positioning code (creating the view and projection matrix) because I thought mine wasn't working well, turns out the maps just poke out behind the camera a bit if the FOV is big enough, so it now works exactly the same but it is mostly LearnOpenGLs implementation. But implemented PCSS, a bigger PCF sampling distance depending on how close/far the occluder is from the point. It performs worse and doesn't really look better, but it is implemented and that's the main thing. PCF just seems to not look that good, it might need TAA to actually make it look fine. But looks even worse at a large sampling radius.