Refactored material sorting/culling
Moved logic into reusable functions. Also changed the sorting to deal with materials when the camera is inside the bounds, should sort so that the ones with the closest farthest away point first. Also changed the behaviour for splitting models into multiple materials is the materials spans too much distance, now takes a minimum faces variable into account and doesn't break a group that would move into groups with less faces, even if it exceeds the bounds, so less groups with 1 triangle so theoretically quicker to sort groups. This hasn't really helped performance noticeably to be honest.