Deep dive into models used for lighting
Used my duplicate project with ImGui hacked in to switch between 2 different models to see the differences real time. In all honesty I had no idea if what I was looking at was realistic or not, but I landed one some models. Most the same or the same but written a bit different from before, but mainly switched the G (geometry masking-shadowing term) to Heitz-Smith GGX. All models used are GGX / Trowbridge-Reitz for NDF, roughness aware Fresnel-Schlick for Fresnel, and Burley for diffuse. Also added a BRDF slice that was rendered with the models/equations from my engine. It displays the general structure it is meant to but a few things that it doesn't have, so would be cool to get a model that has the full correct display, but then again every measured from the real world BRDF slice is different so I shouldn't focus on that too hard.