Adjusted tone mapping
Tried multiple tone maps and decided to go with Reinhard luminance, which aims to keep the colour saturation and adjust only the luminance. I also tried many different exposure and light intensity combos. Ended up with doubling both light intensities (direct and environment) and keeping the exposure to 1, so the raw image definitely has values above 1 which needs the tone mapping. Was very difficult to me, I am so used to the raw, old light intensities image. I may have just adjusted it with the new tone mapping so it looks close to the old image, which is arguably not a correct looking image. I am not an artist. Should probably leave it like this and move on, I could spend ages adjusting various values.