Implemented bloom
Bloom. Hate it, looks bad, doesn't suit game. At a bit of a crossroads because bloom doesn't fit the racing game but I want to continue implementing these effects because I enjoy graphics programming. Have thought about making something that isn't a racing game, maybe a small "open world" thing that you can walk around just while the graphics are expanding, but then I sit down and just want to do work on the graphics instead. SO implemented bloom. Add it to the scene in a composite pass, can pass in more post processing effects into the composite shader and add them to the image accordingly. I can't seem to land on bloom settings that I like, subtle but still there is the aim. But as I said it feels out of place for a racing game, think it would feel more in place for an exploration type game. Also added a blur size, doesn't look good with large blur sizes. To increase the blur size without losing quality you can just pass the texture through the blur shader multiple times, but I wanted to be able to control it on a slider rather than a somewhat discreet "blur one more time and see how it looks", but maybe that's the way to go.