Draw depth into shading pass and use as depth pass
Huge. I tried before drawing the depth pass and blitting (copying as far as I am aware) the depth pass into the shading passes depth buffer so that it could early exit and avoid overdraw based on the depth. Didn't work last time so I just drew the shading pass as normal with depth writing and overdraw. Went back to it and now bind the shading pass and just draw depth into it initially, then use that depth for the SSAO calculations, then just draw colour into the shading pass and this way no overdraws happen. It actually didn't work last time because the skymap was being drawn while depth writing was on when it should have been off, and that was probably also the problem last time (as well as MSAA being on) but it works now. Went from 340fps to 440fps, very cool. Also found an fps gain when not sorting materials, just frustum culling them. No overdraw now means no need to sort them, the whole point in sorting was to minimise overdraw. Still overdraw on the depth texture, but writing depth is so cheap for the gpu and we save so much cpu time on the sorting. Tbh very tiny gain like maybe 3fps on average at 440 but a win is a win.