Adjusted models
If Blender has no haters I am dead. Got rid of clipping (that I thought was me but it was actually built into the model and visible in Blender) on the bridge, removed bushes from the fence visible from second to last corner (you can see the shadows spawn in/not actually be attached because the shadow bias needs to be so big to stop acne), and lightened some textures that were very dark. The model was not worth £15.