Added toggleable bloom and SSAO
Can now turn SSAO or bloom off/on. SSAO gives notable performance increase when off, bloom does not, which is expected. Got annoyed with the z fighting when the near clip plane is close enough that the car doesn't get clipped into in cockpit view. Main culprit is the fence at the bottom of the second to last corner. Had thought that I could eventually add dynamic near/far planes that fit the material tightly, but that gives different values for the depth map which can't have SSAO calculated from it so that won't work. Could have a near pass and a far pass, the near pass renders things close to the camera with a tight far plane, near plane defined by user, then does it's own SSAO, far pass renders everything else with a tight near plane, far plane set by user, and also does it's own SSAO. Getting complicated, not sure of performance implications too but if anything it would be slower, SSAO having to run twice. Nightmare.